requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
beautifulsoup4
python-jose[cryptography]
python-multipart
python-decouple==3.8
gunicorn
SQLAlchemy==2.0.31